update _config,.yml
Getting this site to properly render on Github Pages has been incredibly difficult (especially when it's been rendering fine locally for a day now). This time I'm updating the baseurl to blank and adding the project name to the url. Eventually I hope to find some combination of these settings that will work.